fixed QVT Engine dependencies added test metmamodels

+++ moin/com.sap.tc.moin.repository.test.MoinTestEclipsePlugin/src/com/sap/tc/moin/repository/test/core/TestTypeMorphing.java
@@ -115,7 +115,7 @@ public class TestTypeMorphing extends CoreMoinTest {
         assertTrue( d4sub.isAttributeBoolean( ) );
         assertTrue( d4sub.isAttributeBooleanOptional( ) );
         assertEquals( d4sub.getAttributeDouble( ), 1.1 );
-        assertEquals( d4sub.getAttributeDoubleOptional( ), 1.2 );
+        assertEquals( d4sub.getAttributeDoubleOptional( ), (Double)1.2 );
     }
 
     @Test
